Prevent accidental termination with Ctrl-D
Ctrl-D is used in Google Meet too, to mute and unmute oneself.
Sometimes, the focus is not on the browser, but on the terminal, and a
Bash session gets terminated by accident.

Actually, the shell option `ignoreeof` doesn't disable Ctrl-D
completely; instead, it is equivalent to setting the variable
`IGNOREEOF` to `10`, and thus requires the combination Ctrl-D to be
pressed eleven consecutive times in order to exit the shell session.